+++ reapy/tools/dist_program.py
@@ -1,4 +1,5 @@
 """Define distant Program class."""
+import functools
 
 import reapy
 from reapy.errors import DisabledDistAPIError, DisabledDistAPIWarning
@@ -31,3 +32,25 @@ class Program(program.Program):
             return super(Program, self).run(**input)
         else:
             return CLIENT.run_program(self, input)
+
+    @staticmethod
+    def run_inside(func):
+        """Decorator to make a function/method executable inside Reaper
+            when called from an external app.
+
+            Should only be able to be called from outside Reaper.
+            Parent class' method does not actually decorate `func`.
+
+            TODO: support wrapping property getters/setters if possible.
+        """
+        # check if the decorated function is inside reapy
+        func_module = func.__module__
+        if func_module != 'reapy' and not func_module.startswith('reapy.'):
+            raise RuntimeError('Cannot decorate non-reapy function/method!')
+        # TODO: support decorating non-reapy code some day
+
+        @functools.wraps(func)
+        def _wrap(*args, **kwargs):
+            program = Program(func, None)
+            return program.run(args=args, kwargs=kwargs)
+        return _wrap